# PERF: assorted