--- a/pandas/core/groupby/ops.py
+++ b/pandas/core/groupby/ops.py
@@ -749,7 +749,7 @@ class BaseGrouper:
             # group might be modified
             group_axes = group.axes
             res = f(group)
-            if not _is_indexed_like(res, group_axes, axis):
+            if not mutated and not _is_indexed_like(res, group_axes, axis):
                 mutated = True
             result_values.append(res)