```
import numpy as np
from pandas._libs.internals import BlockPlacement
arr = np.arange(5, dtype=np.intp)

%timeit BlockPlacement(arr)
93.3 ns ± 0.203 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)  # <- PR
1.71 µs ± 40.2 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)  # <- master
```

```
import numpy as np
import pandas as pd

np.random.seed(92342)
arr = np.random.randn(100, 100)
df = pd.DataFrame(arr)
mgr = df._mgr

%timeit mgr._rebuild_blknos_and_blklocs()
6.47 µs ± 132 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)  # <- PR
13.4 µs ± 703 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)  # <- master
```